==> Huan_link_all_script/project/eQTL_Catalogue/script/Knowledge_Graph/script/05_merge_node_markers_annotation.R <==
library(ggplot2)
library(Rcpp)
library(readxl)
library(dplyr)
library(stringr)
library(ggpubr)
library(gridExtra)
library(ggpval)
library(Seurat)

setwd("/home/huanhuan/project/eQTL_Catalogue/script/Knowledge_Graph/output/nodes_annotation/")
hotspot<-read.table("/home/huanhuan/project/eQTL_Catalogue/output/all_tissue_status/hotspot/Tissue_merge_segment_hotspot_cutoff_0.176_extend_18_snp_sorted_merge.bed.gz",header = F,sep = "\t") %>% as.data.frame()
colnames(hotspot) <-c("chr","start","end")
hotspot$Hotspot <-paste(hotspot$chr,hotspot$start,hotspot$end,sep="_")
H3K27ac <- read.table("Adjust_H3K27ac_Tissue_merge_segment_hotspot_cutoff_0.176_extend_18_snp_sorted_merge.bed.gz",header = T,sep = "\t") %>% as.data.frame()
H3K4me1 <- read.table("Adjust_H3K4me1_Tissue_merge_segment_hotspot_cutoff_0.176_extend_18_snp_sorted_merge.bed.gz",header = T,sep = "\t") %>% as.data.frame()
H3K4me3 <- read.table("Adjust_H3K4me3_Tissue_merge_segment_hotspot_cutoff_0.176_extend_18_snp_sorted_merge.bed.gz",header = T,sep = "\t") %>% as.data.frame()
H3K9ac <- read.table("Adjust_H3K9ac_Tissue_merge_segment_hotspot_cutoff_0.176_extend_18_snp_sorted_merge.bed.gz",header = T,sep = "\t") %>% as.data.frame()
H3K36me3 <- read.table("Adjust_H3K36me3_Tissue_merge_segment_hotspot_cutoff_0.176_extend_18_snp_sorted_merge.bed.gz",header = T,sep = "\t") %>% as.data.frame()
H3K27me3 <- read.table("Adjust_H3K27me3_Tissue_merge_segment_hotspot_cutoff_0.176_extend_18_snp_sorted_merge.bed.gz",header = T,sep = "\t") %>% as.data.frame()
H3K9me3 <- read.table("Adjust_H3K9me3_Tissue_merge_segment_hotspot_cutoff_0.176_extend_18_snp_sorted_merge.bed.gz",header = T,sep = "\t") %>% as.data.frame()
CTCF <- read.table("Adjust_CTCF_Tissue_merge_segment_hotspot_cutoff_0.176_extend_18_snp_sorted_merge.bed.gz",header = T,sep = "\t") %>% as.data.frame()
CHROMATIN_Accessibility <- read.table("Adjust_CHROMATIN_Accessibility_Tissue_merge_segment_hotspot_cutoff_0.176_extend_18_snp_sorted_merge.bed.gz",header = T,sep = "\t") %>% as.data.frame()
TFBS <- read.table("Adjust_TFBS_Tissue_merge_segment_hotspot_cutoff_0.176_extend_18_snp_sorted_merge.bed.gz",header = T,sep = "\t") %>% as.data.frame()
Enhancer <- read.table("Adjust_Enhancer_Tissue_merge_segment_hotspot_cutoff_0.176_extend_18_snp_sorted_merge.bed.gz",header = T,sep = "\t") %>% as.data.frame()
eQTL <- read.table("Adjust_hotspot_tissue_anno.bed.gz",header = T,sep = "\t") %>% as.data.frame()

anno <-left_join(hotspot,H3K27ac,by="Hotspot")
anno <-left_join(anno,H3K4me1,by="Hotspot")
anno <-left_join(anno,H3K4me3,by="Hotspot")
anno <-left_join(anno,H3K9ac,by="Hotspot")
anno <-left_join(anno,H3K36me3,by="Hotspot")
anno <-left_join(anno,H3K27me3,by="Hotspot")
anno <-left_join(anno,H3K9me3,by="Hotspot")
anno <-left_join(anno,CTCF,by="Hotspot")
anno <-left_join(anno,CHROMATIN_Accessibility,by="Hotspot")
anno <-left_join(anno,TFBS,by="Hotspot")
anno <-left_join(anno,Enhancer,by="Hotspot")
anno <-left_join(anno,eQTL,by="Hotspot")
write.table(anno,"nodes_marker_enhancers_eqtl_annotation.txt",row.names = F, col.names = T,quote =F,sep="\t")
